Initial commit.
This is a short gpg user guide for beginners. The repository contains
template configuration files for gpg, gpg-agent and dirmngr (keyserver
options). The gpg.sh is a bash script that's must be sourced in e.g.
~/.bashrc if the gpg-agent is used for SSH keys. The script also
exports the environment variable GPG_TTY which must always be defined.
See man page of gpg-agent for details.